move the half custom op
